Moved classes for working with web to another package and changed all required data in other classes
